requirements.txt
# ===========================================
# Pipeline Vélib' - Dépendances Python
# ===========================================

# --- Core Pipeline ---
kafka-python-ng==2.2.3
clickhouse-connect==0.7.19
requests==2.31.0

# --- Monitoring ---
prometheus_client==0.21.0

# --- Testing ---
pytest==8.3.4
pytest-cov==6.0.0

# --- Linting ---
ruff==0.8.6